svn.cpp: Do not call svn_fs_copied_from on delete
Deleted paths will not be in the repository in the revision where they
were deleted. Calling svn_fs_copied_from() on such a path will return an
error and abort operations, which is not what we want here.

Fix this by not calling svn_fs_copied_form if the change is a deletion
(i.e. change->change_kind == svn_fs_path_change_delete).

Failure to do so leads to error messages when exporting a repository:

    Exporting revision 8 svn: E160013: File not found: revision 8, path '/trunk/base/Tcl/pkgIndex.tcl'

Signed-off-by: Clemens Lang <neverpanic@gmail.com>